Split main script for easy maintenance
